Make object fields public & use direct access instead of getters in render loop
Identified the usage of getter methods inside the render loop as one the causes of
performance issues (particularly on older devices) This commit breaks object encapsulation,
but I can't see it causing any real problems as we are unlikely to ever need to modify the
behaviour of the getter. Iideally we would find a better way to handle this.