Trying to improve renderer performance
Preallocate float arrays in each renderer & use buffer.limit() to skip unused values.
The colour float array doesn't need to be allocated for each set of vertices - we can
reuse the same one and change the values. Calls to ComponentFinder methods
are expensive and don't need to be made every render cycle - we can just
store the object references.