Overhaul terrain generation
We can now define a section of the map as a chunk and pass it to the various existing methods to
combine different types of terrain in one map  Adds basic exterior terrain generation. Moved the
cellular automata code into its own class so we can use the code for more than just cavern
generation. Adds a new method for generating indoor rooms (using BSP).