Improve buffer allocation and fix texture binding mistake
Allocating buffers on each frame is expensive so we allocate once initially,
and only reallocate if we need more capacity. We use indicesCount as the
count parameter in glDrawElements to make sure that we only draw the
updated triangles for this frame (vs. drawing the updated triangles +
whatever junk data is in the buffer from previous frame)

Also fixed mistake in code where we were binding the texture on each draw call
(instead of using the already bound textures from ImageLoader)